requirements.txt
alembic==1.3.2
Babel==2.8.0
bcrypt==3.1.7
blinker==1.4
cffi==1.13.2
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Admin==1.5.4
Flask-BabelEx==0.9.3
Flask-Bootstrap4==4.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
passlib==1.7.2
pycparser==2.19
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
six==1.14.0
speaklater==1.3
SQLAlchemy==1.3.12
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
